splits test_tests
